crisis_response_platform/requirements.txt
Flask==2.2.3
Flask-SocketIO==5.3.0
Flask-Login==0.6.2
geopy==2.3.0
python-dotenv==0.19.2
Werkzeug==2.2.3
